pricing strategy added

+++ CarRenting.Host/Entities/RentalAgreement.cs
@@ -7,8 +7,8 @@ namespace CarRenting.Host.Entities
     {
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
-        public decimal RentalPrice { get; set; }
-        public decimal DiscountAmount { get; set; }
+        public double RentalPrice { get; set; }
+        public double DiscountAmount { get; set; }
         public Car RentedCar { get; set; }
         public Customer Customer { get; set; }
     }
